requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
motor==3.4.0
pymongo==4.7.2
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
tweepy==4.14.0
bertopic==0.16.2
sentence-transformers==2.7.0
pandas==2.2.2
python-dotenv==1.0.1
